restructured profile components
