--- frontend/compositions/profile-content/index.tsx
@@ -1,6 +1,0 @@
-import ProfileInfo from "./profile-info"
-import ProfileType from "./profile-type"
-import SavedSearches from "./saved-searches"
-import SavedResults from "./saved-results"
-
-export { ProfileInfo, ProfileType, SavedResults, SavedSearches }
