--- frontend/compositions/profile-content/profile-content.module.css
@@ -1,158 +1,0 @@
-.profileData {
-  padding: var(--size12);
-  max-width: var(--size512);
-}
-
-.sectionTitle {
-  font-size: var(--size24);
-  font-weight: bold;
-  color: var(--darkBlue);
-  margin-bottom: var(--size16);
-}
-
-.row {
-  display: block;
-  line-height: 2;
-}
-.row label {
-  display: block;
-  font-weight: bold;
-  font-size: var(--size16);
-}
-
-.row input {
-  display: block;
-  height: var(--size20);
-}
-
-/* Profile Info */
-
-/* View Mode */
-.editButton {
-  vertical-align: text-top;
-  margin-left: var(--size10);
-  margin-bottom: var(--size12);
-  cursor: pointer;
-}
-
-.dataView {
-  width: 100%;
-  display: flex;
-  flex-flow: row wrap;
-  row-gap: 1rem;
-  column-gap: 2rem;
-}
-
-.dataCell {
-  display: inline-block;
-}
-
-.dataBreak {
-  width: 100%;
-  flex-basis: auto;
-}
-
-.label {
-  display: block;
-  font-size: var(--size10);
-  color: var(--grey);
-  margin-left: 0;
-}
-
-.dataField {
-  font-size: var(--size14);
-  margin-left: var(--size10);
-}
-
-/* Edit Mode */
-
-.profileForm {
-  display: flex;
-  flex-direction: column;
-}
-
-/* copied from "viewer-registration.module.css" */
-.inputLine {
-  display: flex;
-  justify-content: space-evenly;
-}
-
-@media screen and (max-width: 32em) {
-  .inputLine {
-    flex-direction: column;
-    align-items: center;
-  }
-}
-
-.formControls {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
-.formControls button {
-  margin: var(--size16) auto;
-  height: var(--size40);
-  width: var(--size96);
-  border: none;
-  border-radius: var(--size4);
-  font-size: var(--size16);
-  font-weight: 500;
-  cursor: pointer;
-}
-.formControls button:focus,
-.formControls button:hover {
-  -webkit-box-shadow: var(--insetShadow);
-  box-shadow: var(--insetShadow);
-}
-
-.submitButton {
-  background-color: var(--darkBlue);
-  color: white;
-}
-
-.cancelButton {
-  background-color: var(--lightBlue);
-  color: var(--darkBlue);
-}
-
-/* Profile Type */
-
-.profileTypeText {
-  font-size: var(--size36);
-  color: #3473a8;
-  font-weight: bold;
-}
-
-.description {
-  line-height: 2;
-  font-size: var(--size14);
-  color: black;
-  margin-bottom: var(--size20);
-}
-
-.profileTypeLink {
-  font-size: var(--size16);
-  color: var(--linkBlue);
-  text-decoration: underline;
-}
-
-/* Tables */
-
-.savedData {
-  border: 1px solid var(--darkBlue);
-  box-sizing: border-box;
-  box-shadow: -2px 2px 4px 2px rgba(0, 0, 0, 0.25);
-  border-radius: var(--size10);
-}
-
-.dataTitle {
-  font-size: var(--size16);
-  color: var(--lightBlue);
-  line-height: 1.4;
-  display: flex;
-  justify-content: space-between;
-}
-
-.dataGrid {
-  display: block;
-}
