+++ frontend/compositions/profile-info/profile.module.css
@@ -1,0 +1,11 @@
+.profileData {
+  padding: var(--size12);
+  max-width: var(--size512);
+}
+
+.sectionTitle {
+  font-size: var(--size24);
+  font-weight: bold;
+  color: var(--darkBlue);
+  margin-bottom: var(--size16);
+}
