+++ frontend/compositions/profile-info/view-info.module.css
@@ -1,0 +1,48 @@
+.profileData {
+  padding: var(--size12);
+  max-width: var(--size512);
+}
+
+.sectionTitle {
+  font-size: var(--size24);
+  font-weight: bold;
+  color: var(--darkBlue);
+  margin-bottom: var(--size16);
+}
+
+/* View Mode */
+.editButton {
+  vertical-align: text-top;
+  margin-left: var(--size10);
+  margin-bottom: var(--size12);
+  cursor: pointer;
+}
+
+.dataView {
+  width: 100%;
+  display: flex;
+  flex-flow: row wrap;
+  row-gap: 1rem;
+  column-gap: 2rem;
+}
+
+.dataCell {
+  display: inline-block;
+}
+
+.dataBreak {
+  width: 100%;
+  flex-basis: auto;
+}
+
+.label {
+  display: block;
+  font-size: var(--size10);
+  color: var(--grey);
+  margin-left: 0;
+}
+
+.dataField {
+  font-size: var(--size14);
+  margin-left: var(--size10);
+}
